generate database classes
Created a `database` subproject in which the database classes are
generated by applying the Liquibase configuration (also in said
subproject) to the database and then running Slick's code generation.

Source code is currently generated into `target`, should be done into
`src` instead.

I thought that the subproject was needed, but it may be possible to
generate the classes into a single, base project.  Still, it may be
desirable to have them in their own library.